Added check for CCS811 error codes, changed CCS811 data from float to int, do calibrate in the readEnv function

+++ src/main/libraries/EnvData/EnvData.h
@@ -15,8 +15,8 @@ class EnvData {
       float pressure = 0.0;
       float altitude = 0.0;
       float heatIndex = 0.0;
-      float voc = 0.0;
-      float co2 = 0.0;
+      int voc = 0;
+      int co2 = 0;
 };
 
 #endif
